+++ pkgs/test_core/lib/src/runner/reporter/json.dart
@@ -246,13 +246,18 @@ class JsonReporter implements Reporter {
   void _onComplete(LiveTest liveTest) {
     _emit('testDone', {
       'testID': _liveTestIDs[liveTest],
-      // For backwards-compatibility, report skipped tests as successes.
-      'result': liveTest.state.result == Result.skipped
-          ? 'success'
-          : liveTest.state.result.toString(),
+      'result': _normalizeTestResult(liveTest),
       'skipped': liveTest.state.result == Result.skipped,
       'hidden': !_engine.liveTests.contains(liveTest)
     });
+  }
+
+  String _normalizeTestResult(LiveTest liveTest) {
+    // For backwards-compatibility, report skipped tests as successes.
+    if (liveTest.state.result == Result.skipped) return 'success';
+    // if test is still active, it was probably cancelled
+    if (_engine.active.contains(liveTest)) return 'error';
+    return liveTest.state.result.toString();
   }
 
   /// A callback called when [liveTest] throws [error].